Change "c" new-window behaviour
I often prefer the new window to be created after the current one. I suggest to change the default "c" behaviour to reflect this action and to add "C" to create a new window at the end of the list.
It's easy to think of "c" as "light action" (create here) and "C" as "strong action" (create there, push the new window far).